clase5/app/requirements.txt
Flask== 2.0.2 
mysql-connector-python==8.0.32 
pymysql==1.0.2 
Werkzeug==2.1.2